+++ src/app/shared/change-pass/change-pass.component.ts
@@ -10,6 +10,7 @@ import { CONFIG } from '../../../../config';
 import { ToastrService } from 'ngx-toastr';
 import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
+import { NetworkService } from '../../service/network.service';
 
 @Component({
   selector: 'app-change-pass',
@@ -19,6 +20,7 @@ import { CommonModule } from '@angular/common';
 })
 export class ChangePassComponent {
   changePassword!: FormGroup;
+  loader: boolean = false
   confirmPasswordMatch: boolean = false;
   isOld: boolean = false;
   isNew: boolean = false;
@@ -31,7 +33,8 @@ export class ChangePassComponent {
     private fb: FormBuilder,
     private http: HttpClient,
     private toaster: ToastrService,
-    private router: Router
+    private router: Router,
+    private appService: NetworkService
   ) {
     this.changePassword = this.fb.group({
       newPassword: ['', Validators.required],
@@ -85,33 +88,22 @@ export class ChangePassComponent {
       newPassword: this.changePassword.get('newPassword')?.value,
     };
 
-    if (this.changePassword.valid) {
+    if (this.changePassword.valid && !this.loader) {
+      this.loader = true
       this.http.post(CONFIG.userChangePassword, req).subscribe({
         next: (res: any) => {
           this.toaster.success(res.meta.message);
           localStorage.clear();
           this.router.navigate(['/']);
+          this.loader = false
           this.changePassword.reset();
         },
 
         error: (error: any) => {
-          let errorObject = error.meta
+          this.appService.ErrorNotification_Manager(error.error);
+          this.loader = false
             ? error.meta.message
             : error.error?.meta?.message;
-          if (typeof errorObject === 'object') {
-            for (var key of Object.keys(errorObject)) {
-              this.toaster.error(errorObject[key].message, '', {
-                positionClass: 'toast-top-right',
-              });
-              return;
-            }
-          } else {
-            this.toaster.error(errorObject, '', {
-              positionClass: 'toast-top-right',
-              timeOut: 800,
-            });
-            return;
-          }
         },
       });
     }
